Kudos! We have a working searchByValue method. Now -- the next part is to make search methods caseInsensitive. I'm goping to answer the prompts in the problem here.
Which methods are called when searching?
PrintJobs
SearchByValue
(possibly search by column and value? it would depend)
How is the user’s search string compared against the values of fields of the job HashMap objects?

How can you make this comparison in a way that effectively ignores the case of the strings?

How can you do this without altering the capitalization of the items in allJobs so that the data gets printed out the same way that it appears in job_data.csv?

Nope. I think this is really simple